Merge into 1 crate
The three files in hidpipe are very much interconnected and form a
single unit split into two binaries. Splitting that into 3 crates (==> 3
packages) is unnecessarily complex. It's easier to treat hidpipe as a
single unit - which it is, just has two binaries of the same unit for
each side of the VM boundary. That in turn means 1 fedora package
instead of 3. It also means 1 cargo.toml instead of 4. Which is a
significant savings for a project of this small size.

This also corrects the licensing issue so is an alternative to https://github.com/WhatAmISupposedToPutHere/hidpipe/pull/1

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>